Ready for pull request. Remaining issues:
Error messages containing timestamps and/or object references cause mis-matches between expected and real values and therefore cause problems. Some ad-hoc fixes are in place, but the problems still persist and an elegant solution is still required.

Proxyprint (is the current example we have, but it's an interesting problem):
Type definitions appear to vary, for specialist JSON types, between the type (as string) that is found in EvoMaster and the same type (as string) obtained from RestAssured. More work on this.

UPDATE: added some stuff for accept() to address the ProxyPrint problems. It solves some (but not all).