Changes related to pull request. Refactored mode to be an enum. Removed unnecessary mutable lists, and replaced with list (had missed a couple before, sorry about that).
Changed AbstractRestFitness.handleRestCall to use the GeneUtils.applyEscapes
(to keep a centralized handling of escapes).
RestPath.resolveOnlyPath seems to use only raw string values,
RestPath.encode uses URLEncoder. Not sure if escapes need to be handled separately here.

TestSuiteOrganizer isn't meant to be static. Not sure if I missed something.
Alternatively, the helpers can be made stateless (the only stateful thing is
the currently selected criteria for each. That is more of a "we might need it later"
type of convenience that we can dispense with, if needed.