WiP on expectations.
Getting model information on response objects to the writer.
Getting Model information to the writer.
Next step: expect that returned objects conform to the stated object
structure. (E.g. if the object definition has required fields, those
fields are present in the returned object)
This will be added as a first in a series of partial oracles that will
drive the expectation generation.