WiP. Filtering out "logged" (to avoid ID troubles).
expectedRefs is now a map that also stores the code where
each particular ref is expected.

Checking that structure is only assessed for codes that should
expect some object (i.e. not 500s, or variations of 400).

Known issue: when a MapGene is processed, it's name gets a "_map" added
to its tail. This causes problems when comparing expected keys vs
present keys.