WiP. Clustering seems to work, in principle.
The main parameter is epsilon (the accepted distance within clusters).

Due to changes in how the distance is computed (Levenshtein distance
divided by the length of the larger string), the epsilon is expected
to be between 0.0 and 1.0. (note that nothing actually stops one from
using larger values, just the results will be wonky).

More experimentation is needed to figure out good values for epsilon,
but 0.5 or 0.6 seem to work for now. This implies that results will be
in the same cluster if the error messages (only clustering 500s and 400s)
are about 50% similar. This is to account for the same error message
making reference to different dates/timestamps/objects/etc.