Initial version of schema oracle.
Note: Since the current SUTs have auto-generated schemas, none of these
should cause problems. But a spot of refactoring will be needed once
new SUTs are in place.

Note 2: There is a (seemingly known) bug in Groovy regarding the
handling of JSON that only has basic strings. Right now that does not
generate an expectation but it may be worth keeping an eye on in future.