Updated expectation generation for the SchemaOracle.kt - now also includes generating expectations on arrays of objects.
The code expects that all the objects in the array are compatible in structure with the expected objects (as defined by the OpenAPI).

Note: comparing types is somewhat tricky, as the response does not contain information about the type (so two objects with different types but compatible structures would be considered to match).