fix: solved version of the process monitor bug.
I commented out the "evalInd != eval" assignment because the changes made in the "getMutatator().mutateAndSave(nMutations, ei, archive)" line is not reported to processMonitor. This may be a side-effect. It should be double-checked.

Also, the ProcessMonitorTest could not catch the StackOverflow exception in the ProcessMonitor. A new test was added.