AI models generalization
The updated `AIGaussianCheck` and `AIGLMCheck` demonstrate that the models are compatible with any API. If the API includes complex endpoints with unsupported genes, the classifier will return `null`, and testing will fall back to random sampling.

The petShop API has been updated to include such an endpoint for validation purposes.

In `AIResponseClassification`, the term `probability` has been replaced by `score`, which is a more general term representing the likelihood of occurrence depending on the selected model.